requirements.txt
langchain-text-splitters
langchain-community
langchain-core
langchain-openai
langchain-pinecone
langgraph
bs4
python-dotenv
pinecone